Rework match and team commands.
This is a pretty significant change to the way data was being saved. I
have not done anything with commands that were not in the match cog, but
the match cog no longer uses transactionConfiguration to help it load or
store data. Instead, it uses a new rscdata cog. Data files are now
created per "dataset", which is currently separated by cog. So, match
has a file in the rscdata directory. The new cog teamManager also has a
file there. The teamManager cog is now where teamList, tierList,
addTier, and removeTier live.